requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
torch==2.0.0
torchvision==0.15.0
Pillow==9.3.0
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
typing-extensions>=4.0.0
loguru==0.7.0